Implementado el endpoint de listar administradores, el crud de estudiantes y grupos

+++ src/main/java/com/sistemas_mangager_be/edu_virtual_ufps/repositories/MatriculaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sistemas_mangager_be.edu_virtual_ufps.entities.Matricula;
 
-public interface MatriculaRepository extends JpaRepository <Matricula, Integer> {
+public interface MatriculaRepository extends JpaRepository <Matricula, Long> {
     
 }
